Customize profile_message_execution task from template
The previous commit added the profile_message_execution task based on a template by running the add_task.sh script. This commit provides customizations to the initial template.

The output handling and data retrieval method of the run bash script has been modified. Instead of redirecting the output to a new file, the script is now executed directly via SSH. Additionally, instead of outputting to a single file, a new directory is created based on the current date and time, and two profiling SVGs are copied from the remote server to this new directory using SCP.

The run_task bash script, rather than echoing a placeholder message, now executes the lotus daemon with a ten-minute time limit. Two different pprof commands are run to generate detailed profiling data. These modifications enhance the quality and granularity of the profiling data generated by the task.